+++ NINA.Profile/Interfaces/ICameraSettings.cs
@@ -63,5 +63,10 @@ namespace NINA.Profile.Interfaces {
         ushort? AtikGainPreset { get; set; }
         ushort? AtikExposureSpeed { get; set; }
         int? AtikWindowHeaterPowerLevel { get; set; }
+
+        bool TouptekAlikeHighGain { get; set; }
+        bool TouptekAlikeUltraMode { get; set; }
+        bool TouptekAlikeHighFullwell { get; set; }
+        int TouptekAlikeDewHeaterStrength { get; set; }
     }
 }